Make sure all meshlet faces come from a single geometry.
Unique meshlet geometry index significantly simplifies meshlet storage and rendering. Quality of decimation is maintained at approximately the same level because meshlet groups might still contain meshlets from different geometries.

Usability improvements:
- There is no need for a variable size geometry_index lookup table per meshlet.
- During rendering it is possible to bin meshlets by material index or material properties (two sided, dithered, alpha tested, etc).

With this change KD tree lookups visit 3-4 times more nodes (for 2 geometries per mesh) which results in a slower meshlet construction. Potential performance improvements:
- Add node partitioning by geometry index to move the filtering higher up the tree and improve culling efficiency.
- Store a per node mask of geometries (1u << (hash(geometry_index) % 32)) and use it for conservative culling.

+++ main.cpp
@@ -213,7 +213,7 @@ int main() {
 	t0 = std::chrono::high_resolution_clock::now();
 	
 #if 0
-	u32 split_index = ((triangle_mesh.indices.size() / 3) / 2) * 3;
+	u32 split_index = (((u32)triangle_mesh.indices.size() / 3) / 2) * 3;
 	
 	compile_const u32 geometry_desc_count = 2;
 	TriangleGeometryDesc geometry_descs[geometry_desc_count];
